♻ Refactored code a bit.
Moved some of the functions that aren't really part of the asyncAirtable class to their own separate files.
