+++ lib/asyncAirtable.js
@@ -1,4 +1,7 @@
 const nodeFetch = require('node-fetch');
+const buildOpts = require('./buildOpts');
+const checkError = require('./checkError');
+const checkArg = require('./checkArg');
 const baseURL = 'https://api.airtable.com/v0';
 
 const validOptions = [
@@ -11,6 +14,8 @@ const validOptions = [
 ];
 
 class AsyncAirtable {
+  rateLimited = false;
+
   constructor(apiKey, base) {
     if (!apiKey) throw new Error('API Key is required.');
     if (!base) throw new Error('Base ID is required.');
@@ -18,50 +23,11 @@ class AsyncAirtable {
     this.base = base;
   }
 
-  buildOpts = (opts) => {
-    const params = Object.keys(opts)
-      .map((key, i) => {
-        const opt = opts[key];
-        let formatted;
-        if (Array.isArray(opt)) {
-          formatted = opt
-            .map((item, j) => {
-              switch (typeof item) {
-                case 'object':
-                  return Object.keys(item)
-                    .map((subKey) => {
-                      return `${key}[${j}][${subKey}]=${item[subKey]}`;
-                    })
-                    .join('&');
-                case 'string':
-                  return `${key}[]=${item}`;
-              }
-            })
-            .join('&');
-        } else {
-          formatted = `${key}=${opt}`;
-        }
-        return i !== 0 ? `&${formatted}` : formatted;
-      })
-      .join('');
-    return encodeURI(params);
-  };
-
-  checkError = (status) => !(status < 300 && status >= 200);
-
-  checkArg = (arg, name, type, required) => {
-    if (!arg && required) throw new Error(`Argument "${name}" is required.`);
-    if (arg && typeof arg !== type)
-      throw new Error(
-        `Incorrect data type on argument "${name}". Received "${typeof arg}": expected "${type}"`,
-      );
-  };
-
   select = async (table, options, page) => {
     try {
-      this.checkArg(table, 'table', 'string', true);
-      this.checkArg(options, 'options', 'object');
-      this.checkArg(page, 'page', 'number');
+      checkArg(table, 'table', 'string', true);
+      checkArg(options, 'options', 'object');
+      checkArg(page, 'page', 'number');
       let url = `${baseURL}/${this.base}/${table}`;
       const opts = options ? { ...options } : {};
       Object.keys(opts).forEach((option) => {
@@ -76,11 +42,11 @@ class AsyncAirtable {
             opts.offset = offset;
           }
           try {
-            const res = await nodeFetch(`${url}?${this.buildOpts(opts)}`, {
+            const res = await nodeFetch(`${url}?${buildOpts(opts)}`, {
               headers: { Authorization: `Bearer ${this.apiKey}` },
             });
             const body = await res.json();
-            if (this.checkError(res.status)) {
+            if (checkError(res.status)) {
               throw new Error(JSON.stringify(body));
             }
             if (i + 1 === page) {
@@ -99,11 +65,11 @@ class AsyncAirtable {
             opts.offset = offset;
           }
           try {
-            const res = await nodeFetch(`${url}?${this.buildOpts(opts)}`, {
+            const res = await nodeFetch(`${url}?${buildOpts(opts)}`, {
               headers: { Authorization: `Bearer ${this.apiKey}` },
             });
             const body = await res.json();
-            if (this.checkError(res.status)) {
+            if (checkError(res.status)) {
               throw new Error(JSON.stringify(body));
             }
             data = data.concat(body.records);
@@ -124,14 +90,14 @@ class AsyncAirtable {
 
   find = async (table, id) => {
     try {
-      this.checkArg(table, 'table', 'string', true);
-      this.checkArg(id, 'id', 'string', true);
+      checkArg(table, 'table', 'string', true);
+      checkArg(id, 'id', 'string', true);
       let url = `${baseURL}/${this.base}/${table}/${id}`;
       const res = await nodeFetch(url, {
         headers: { Authorization: `Bearer ${this.apiKey}` },
       });
       const data = await res.json();
-      if (this.checkError(res.status)) {
+      if (checkError(res.status)) {
         throw new Error(JSON.stringify(data));
       }
       return data;
@@ -142,8 +108,8 @@ class AsyncAirtable {
 
   createRecord = async (table, record) => {
     try {
-      this.checkArg(table, 'table', 'string', true);
-      this.checkArg(record, 'record', 'object', true);
+      checkArg(table, 'table', 'string', true);
+      checkArg(record, 'record', 'object', true);
       let url = `${baseURL}/${this.base}/${table}`;
       const body = { fields: record };
       const res = await nodeFetch(url, {
@@ -155,7 +121,7 @@ class AsyncAirtable {
         },
       });
       const data = await res.json();
-      if (this.checkError(res.status)) {
+      if (checkError(res.status)) {
         throw new Error(JSON.stringify(data));
       }
       return data;
@@ -166,8 +132,8 @@ class AsyncAirtable {
 
   updateRecord = async (table, record, destructive = false) => {
     try {
-      this.checkArg(table, 'table', 'string', true);
-      this.checkArg(record, 'record', 'object', true);
+      checkArg(table, 'table', 'string', true);
+      checkArg(record, 'record', 'object', true);
       let url = `${baseURL}/${this.base}/${table}/${record.id}`;
       const fields = {};
       Object.keys(record).forEach((key) => {
@@ -182,7 +148,7 @@ class AsyncAirtable {
         },
       });
       const data = await res.json();
-      if (this.checkError(res.status)) {
+      if (checkError(res.status)) {
         throw new Error(JSON.stringify(data));
       }
       return data;
@@ -193,8 +159,8 @@ class AsyncAirtable {
 
   deleteRecord = async (table, id) => {
     try {
-      this.checkArg(table, 'table', 'string', true);
-      this.checkArg(id, 'id', 'string', true);
+      checkArg(table, 'table', 'string', true);
+      checkArg(id, 'id', 'string', true);
       let url = `${baseURL}/${this.base}/${table}/${id}`;
       const res = await nodeFetch(url, {
         method: 'delete',
@@ -203,7 +169,7 @@ class AsyncAirtable {
         },
       });
       const data = await res.json();
-      if (this.checkError(res.status)) {
+      if (checkError(res.status)) {
         throw new Error(JSON.stringify(data));
       }
       return data;
@@ -214,8 +180,8 @@ class AsyncAirtable {
 
   bulkCreate = async (table, records) => {
     try {
-      this.checkArg(table, 'table', 'string', true);
-      this.checkArg(records, 'records', 'object', true);
+      checkArg(table, 'table', 'string', true);
+      checkArg(records, 'records', 'object', true);
       let url = `${baseURL}/${this.base}/${table}`;
       const body = records.map((record) => ({
         fields: record,
@@ -229,7 +195,7 @@ class AsyncAirtable {
         },
       });
       const data = await res.json();
-      if (this.checkError(res.status)) {
+      if (checkError(res.status)) {
         throw new Error(JSON.stringify(data));
       }
       return data.records;
@@ -240,8 +206,8 @@ class AsyncAirtable {
 
   bulkUpdate = async (table, records) => {
     try {
-      this.checkArg(table, 'table', 'string', true);
-      this.checkArg(records, 'records', 'object', true);
+      checkArg(table, 'table', 'string', true);
+      checkArg(records, 'records', 'object', true);
       let url = `${baseURL}/${this.base}/${table}`;
       const body = records.map((record) => {
         const id = record.id;
@@ -260,7 +226,7 @@ class AsyncAirtable {
         },
       });
       const data = await res.json();
-      if (this.checkError(res.status)) {
+      if (checkError(res.status)) {
         throw new Error(JSON.stringify(data));
       }
       return data.records;
@@ -271,8 +237,8 @@ class AsyncAirtable {
 
   bulkDelete = async (table, ids) => {
     try {
-      this.checkArg(table, 'table', 'string', true);
-      this.checkArg(ids, 'ids', 'object', true);
+      checkArg(table, 'table', 'string', true);
+      checkArg(ids, 'ids', 'object', true);
       let query = '';
       ids.forEach((id, i) => {
         if (i !== 0) {
@@ -289,7 +255,7 @@ class AsyncAirtable {
         },
       });
       const data = await res.json();
-      if (this.checkError(res.status)) {
+      if (checkError(res.status)) {
         throw new Error(JSON.stringify(data));
       }
       return data.records;
