+++ lib/checkError.js
@@ -1,0 +1,1 @@
+module.exports = (status) => !(status < 300 && status >= 200);
